% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_heart_data.R
\name{clean_heart_data}
\alias{clean_heart_data}
\title{Clean Heart Attack Dataset}
\usage{
clean_heart_data(df)
}
\arguments{
\item{df}{A data frame of raw heart attack data.}
}
\value{
A cleaned data frame with selected columns and categorical columns converted to factors.
}
\description{
Removes unnecessary columns and converts relevant columns to factors.
}
