% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_significant_variables.R
\name{get_significant_variables}
\alias{get_significant_variables}
\title{Get significant variables from logistic regression model}
\usage{
get_significant_variables(data, target = "Heart_Attack_Risk", p_thresh = 0.15)
}
\arguments{
\item{data}{A data frame with predictor variables and a binary target variable.}

\item{target}{Name of the binary outcome column (as a string).}

\item{p_thresh}{Significance threshold for p-values (default = 0.15)}
}
\value{
A character vector of significant variable names.
}
\description{
This function trains a logistic regression model and returns the names of variables
whose p-values are below a specified threshold.
}
